11/28 Change DiaryList.kt to WriteDiary.kt, and create diary_list.xml and arrays.xml
