+++ MyDiary/app/src/main/AndroidManifest.xml
@@ -15,6 +15,9 @@
         <activity android:name=".EmotionSe"/>
         <activity android:name=".ViewPage"/>
         <activity android:name=".CreateAccount"/>
+        <activity android:name=".WriteDiary" android:theme="@style/Theme.AppCompat.Light.NoActionBar"/>
+        <activity android:name=".DiaryList"/>
+        <activity android:name=".PutPicture"/>
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
@@ -25,11 +28,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".DiaryList"
-            android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-
-        </activity>
-        <activity android:name=".PutPicture"></activity>
     </application>
 
 </manifest>
